+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ## Changelog for Gmail Auto-forwarding
+- **Version 1.3 build 20240624**
+```
+- Fixed forwarded emails are sometimes not deleted due to the Google service not responding
+- Forwarded email sent as an attachment as the workaround for body size limitations and to
+  keep the forwarded email format intact
+- Forwarded email sent per recipient
+- Add new advanced features (Advanced user only):
+  - Add a new feature to forward email using a registered email alias
+  - Add a new feature to wait for the running execution to end first to prevent overlapping execution
+```
 - **Version 1.2 build 20240329**
 ```
 - Add a new feature to validate the user's input